Extracted ResourceSetValidation Listener and SCTValidationJob is a Singleton binding now to avoid multiple validations in case of subdiagrams
